assessement/requirements.txt
requests 
djangorestframework
Django
gunicorn
psycopg2-binary 
schedule
pytest
django-debug-toolbar
django-redis



